Converted racer_thread type definition into opaque pointer.
